[Fix] 옵션 에러 수정
