--- Assets/External/Infima Games/Low Poly Shooter Pack/Audio/Mixers/AM_Default.mixer
@@ -1,65 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!241 &24100000
-AudioMixerController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: AM_Default
-  m_OutputGroup: {fileID: 0}
-  m_MasterGroup: {fileID: 24300002}
-  m_Snapshots:
-  - {fileID: 24500006}
-  m_StartSnapshot: {fileID: 24500006}
-  m_SuspendThreshold: -80
-  m_EnableSuspend: 1
-  m_UpdateMode: 0
-  m_ExposedParameters: []
-  m_AudioMixerGroupViews: []
-  m_CurrentViewIndex: 0
-  m_TargetSnapshot: {fileID: 24500006}
---- !u!243 &24300002
-AudioMixerGroupController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: Master
-  m_AudioMixer: {fileID: 24100000}
-  m_GroupID: 89ac416638507be45a2fd0c6132e68b7
-  m_Children: []
-  m_Volume: b6494b71360b629439b277851f4c2d10
-  m_Pitch: 352b6b097baee204aab238891c39998a
-  m_Send: 00000000000000000000000000000000
-  m_Effects:
-  - {fileID: 24400004}
-  m_UserColorIndex: 0
-  m_Mute: 0
-  m_Solo: 0
-  m_BypassEffects: 0
---- !u!244 &24400004
-AudioMixerEffectController:
-  m_ObjectHideFlags: 3
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: 
-  m_EffectID: e9a0f402742f70140bb9d11492559e01
-  m_EffectName: Attenuation
-  m_MixLevel: a4d63af498e17374cb7b62928de1c505
-  m_Parameters: []
-  m_SendTarget: {fileID: 0}
-  m_EnableWetMix: 0
-  m_Bypass: 0
---- !u!245 &24500006
-AudioMixerSnapshotController:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: Snapshot
-  m_AudioMixer: {fileID: 24100000}
-  m_SnapshotID: e1621fbc7cbab3a4dab1ffdc0b3463cc
-  m_FloatValues: {}
-  m_TransitionOverrides: {}
