+++ Assets/Player.mixer.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3fcdb23bde063e94fa0b60ccf87604d6
+guid: e72f2871aae218d4aa2e3dc469ff9c54
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 24100000
