+++ Assets/Scripts/Ui/MouseSlider.cs
@@ -3,11 +3,27 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class MouseSlider : SliderValue
+public class MouseSlider : MonoBehaviour
 {
-    public override void Apply()
+    [SerializeField]
+    private static float saved = 50;
+    [SerializeField]
+    private float max = 2;
+    private Slider slider;
+
+    void Start()
     {
-        base.Apply();
+        slider = GetComponent<Slider>();
+        slider.value = saved;
+    }
+
+    public void Cancel()
+    {
+        slider.value = saved;
+    }
+    public void Apply()
+    {
+        saved = slider.value;
         MouseControl.rotCamXAxisSpeed = saved / 100 * max;
         MouseControl.rotCamYAxisSpeed = saved / 100 * max;
     }
