--- Assets/Scripts/Ui/SliderValue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d5d3b0f4112594498e19ba16c0c3906
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
